Add wildcard constraint to readgroup
The lack of a wildcard constraint would trigger a bug when the input
files matched the same pattern as the output from cutadapt
(*_R1.fastq.gz). Since the readgroup is generated by the pipeline, we
know it always confirms to regex 'rg\d+', so we can add this constraint
safely, without limiting the user in their choice of sample name.